GET query implemented for Students Blog Section
